feat: 後台Spec相關API移至`AdminControllers`資料夾內 1. 新增`AdminSpecController`

+++ ApplicationCore/Services/SpecService.cs
@@ -37,43 +37,46 @@ namespace ApplicationCore.Services
             catch (Exception ex) 
             {
                 _logger.LogError(ex, ex.Message);
-                return new OperationResult("產品規格新增失敗");
+                return new OperationResult("商品規格新增失敗");
             }
         }
 
         public async Task<OperationResult> UpdateSpecAsync(UpdateSpecDTO updateSpecDTO)
         {
-            var spec = await _specRepository.GetByIdAsync(updateSpecDTO.SpecId);
-
-            spec.Sku = updateSpecDTO.SKU;
-            spec.UnitPrice = updateSpecDTO.UnitPrice;
-            spec.SpecName = updateSpecDTO.SpecName;
-            spec.Order = updateSpecDTO.Order;
-            spec.StockQuantity = updateSpecDTO.StockQuantity;
-
             try
             {
+                var spec = await _specRepository.GetByIdAsync(updateSpecDTO.SpecId);
+                if (spec == null) return new OperationResult("找不到對應的商品規格");
+
+                spec.Sku = updateSpecDTO.SKU;
+                spec.UnitPrice = updateSpecDTO.UnitPrice;
+                spec.SpecName = updateSpecDTO.SpecName;
+                spec.Order = updateSpecDTO.Order;
+                spec.StockQuantity = updateSpecDTO.StockQuantity;
+
                 await _specRepository.UpdateAsync(spec);
                 return new OperationResult();
             }
             catch(Exception ex)
             {
-                return new OperationResult("產品規格更新失敗");
+                return new OperationResult("商品規格更新失敗");
             }
         }
 
         public async Task<OperationResult> DeleteSpecAsync(int specId)
         {
-            var spec = await _specRepository.GetByIdAsync(specId);
             try
             {
+                var spec = await _specRepository.GetByIdAsync(specId);
+                if (spec == null) return new OperationResult("找不到對應的商品規格");
+
                 await _specRepository.DeleteAsync(spec);
                 return new OperationResult();
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex,ex.Message);
-                return new OperationResult("產品規格刪除失敗");
+                return new OperationResult("商品規格刪除失敗");
             }
         }
     }
